requirements.txt
pyyaml>=6.0
requests>=2.28
pytz>=2023.3
